% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_bicluster.R
\name{col_spectral}
\alias{col_spectral}
\title{spectral clustering for the column vectors}
\usage{
col_spectral(X, k = 5)
}
\arguments{
\item{X}{noised matrix}

\item{k}{the number of clusters}
}
\value{
cluster: the cluster assignment for column vectors
}
\description{
spectral clustering for the column vectors
}
