% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx_bicluster.R
\name{cvx_bicluster}
\alias{cvx_bicluster}
\title{Use convex bicluster algorithm to detect the row pattern and column pattern}
\usage{
cvx_bicluster(X, X_baseline, k_row = 5, k_col = 5, phi = 0.5, nGamma = 10)
}
\arguments{
\item{X}{the matrix waited to be biclustered}

\item{X_baseline}{the groundtruth to compute the clustering accuracy}
}
\value{
clu_row: cluster of row vectors

clu_col: cluster of column vectors

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
Use convex bicluster algorithm to detect the row pattern and column pattern
}
