% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx_bicluster.R
\name{cvx_cluster}
\alias{cvx_cluster}
\title{Detect the cluster pattern in row and column direction}
\usage{
cvx_cluster(X, k = 5, phi = 0.5, gamma)
}
\arguments{
\item{X}{the matrix waited to be biclustered}

\item{X_baseline}{the groundtruth to compute the clustering accuracy}
}
\value{
clu_row: cluster of row vectors

clu_col: cluster of column vectors

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
Detect the cluster pattern in row and column direction
}
