% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvx_bicluster.R
\name{cvx_separate_cluster}
\alias{cvx_separate_cluster}
\title{Use convex clustering algorithm to separately detect the row pattern and column pattern}
\usage{
cvx_separate_cluster(
  X,
  X_baseline,
  row_k = 5,
  col_k = 5,
  phi = 0.5,
  row_gamma = 5,
  col_gamma = 5
)
}
\arguments{
\item{X}{the matrix waited to be biclustered}

\item{X_baseline}{the groundtruth to compute the clustering accuracy}
}
\value{
clu_row: cluster of row vectors

clu_col: cluster of column vectors

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
Use convex clustering algorithm to separately detect the row pattern and column pattern
}
