% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_bicluster.R
\name{kl_cluster}
\alias{kl_cluster}
\title{objective: Use KL divergence based k means clustering to cluster distributions.}
\usage{
kl_cluster(mu, cov, k = 5, max_it = 1000)
}
\arguments{
\item{mu}{a m x n matrix where each row is a n-dimensional mean vector of a distribution}

\item{cov}{a list of m n x n matrices where each matrix is a covariance matrix 
of a distribution}

\item{k}{number of clusters}
}
\value{
clusters: a group of clusters c_1, c_2,...c_k where c_i is the ith cluster of distributions

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
objective: Use KL divergence based k means clustering to cluster distributions.
}
