% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_bicluster.R
\name{kmeans_multiple_bicluster}
\alias{kmeans_multiple_bicluster}
\title{Apply k means clustering based biclustering algorithm to detect the row and column partitions}
\usage{
kmeans_multiple_bicluster(X, X_baseline, row_k = 5, col_k = 5, type = "KL")
}
\arguments{
\item{X}{noised matrix}

\item{X_baseline}{the ground truth to compute the clustering accuracy}

\item{row_k}{the number of clusters in row direction}

\item{col_k}{the number of clusters in column direction}

\item{type}{the clusterin algorithms used, 'KL' represents the KL divergence, 'mean_vecotr' represents the traditional k means}
}
\value{
clu_row: cluster of row vectors

clu_col: cluster of column vectors

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
Apply k means clustering based biclustering algorithm to detect the row and column partitions
}
