% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_bicluster.R
\name{kmeans_multiple_cluster}
\alias{kmeans_multiple_cluster}
\title{Achieve k means multiple sample clustering in one dimension based on the clustering information of another dimension.}
\usage{
kmeans_multiple_cluster(X, partition, k = 5, flag = "row", type = "KL")
}
\arguments{
\item{X}{a m x n noised matrix with additive Gaussian noise. Elements of a row vectors belongs to k clusters.
The elements in the same cluster can be view as a i.i.d sample of a hidden distribution.
These k distributions are independent}

\item{partition}{the clustering information of columns}

\item{k}{number of clusters in row/column direction}

\item{flag}{can be "row" or "col", achieve row or column clustering based on the value of flag}

\item{type}{choose the corresponded multiple sample clustering algorithm}
}
\value{
cluster: the cluster for column/row vectors
}
\description{
Achieve k means multiple sample clustering in one dimension based on the clustering information of another dimension.
}
