% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_bicluster.R
\name{kmeans_separate_bicluster}
\alias{kmeans_separate_bicluster}
\title{Apply k means clustering based separated biclustering algorithm to detect the row and column partitions}
\usage{
kmeans_separate_bicluster(X, X_baseline, row_k = 5, col_k = 5)
}
\arguments{
\item{X}{noised matrix}

\item{X_baseline}{the ground truth to compute the clustering accuracy}

\item{row_k}{the number of clusters in row direction}

\item{col_k}{the number of clusters in column direction}
}
\value{
clu_row: cluster of row vectors

clu_col: cluster of column vectors

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
Apply k means clustering based separated biclustering algorithm to detect the row and column partitions
}
