% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_bicluster.R
\name{mean_vector}
\alias{mean_vector}
\title{objective: compute wasserstain distance between two Gaussian distributions}
\usage{
mean_vector(mu, k = 5, weight = rep(x = 1, times = ncol(mu)))
}
\arguments{
\item{mu1}{the mean vectors of distribution1}

\item{mu2}{the mean vectors of distribution2}

\item{cov1}{the covariance matrix of distribution1}

\item{cov2}{the covariance matrix of distribution2}
}
\value{
Wasserstain distance between two distributions
}
\description{
objective: compute wasserstain distance between two Gaussian distributions
}
