% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_bicluster.R
\name{row_spectral}
\alias{row_spectral}
\title{spectral clustering for the row vectors}
\usage{
row_spectral(X, k = 5)
}
\arguments{
\item{X}{noised matrix}

\item{k}{the number of clusters}
}
\value{
cluster: the cluster assignment for row vectors
}
\description{
spectral clustering for the row vectors
}
