% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_bicluster.R
\name{spectral_separate_bicluster}
\alias{spectral_separate_bicluster}
\title{Separately apply spectral clustering on the column and row vectors to detect the row and column partitions}
\usage{
spectral_separate_bicluster(X, X_baseline, row_k = 5, col_k = 5)
}
\arguments{
\item{X}{noised matrix}

\item{X_baseline}{the ground truth to compute the clustering accuracy}

\item{row_k}{the number of clusters in row direction}

\item{col_k}{the number of clusters in column direction}
}
\value{
clu_row: cluster of row vectors

clu_col: cluster of column vectors

RI: rand index to measure the clustering accuracy

ARI: adjusted rand index to measure the clustering accuracy

VI: variance of information to measure the clustering accuracy
}
\description{
Separately apply spectral clustering on the column and row vectors to detect the row and column partitions
}
