% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_sep_bicluster.R
\name{svd_separated_bicluster}
\alias{svd_separated_bicluster}
\title{Use svd to separate the clustering information and separately detect the clustering information}
\usage{
svd_separated_bicluster(
  X,
  X_baseline,
  phi = 0.5,
  k = 5,
  nGamma = 5,
  r = 3,
  method = "admm",
  flag = 0
)
}
\arguments{
\item{X}{m x n matrix wait to be clustered}

\item{X_baseline}{the ground truth of matrix X}

\item{phi}{weight that controls the penalty term of checkerboard pattern}

\item{k}{hyperparameter for k-nearest neighborhood}

\item{nGamma}{how many regularization parameters are tried}

\item{r}{control the rank of singular matrix}

\item{method}{Algorithm to use: "admm" or "ama"}

\item{flag}{if auto-tuning parameter. 0 -> off, 1 -> on}

\item{cluster}{cluster of row vectors}

\item{RI}{rand index to measure the clustering accuracy}

\item{ARI}{adjusted rand index to measure the clustering accuracy}

\item{VI}{variance of information to measure the clustering accuracy}
}
\description{
Input:
}
