% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_bicluster.R
\name{wasserstain_compute}
\alias{wasserstain_compute}
\title{objective: compute wasserstain distance between two Gaussian distributions}
\usage{
wasserstain_compute(mu1, mu2, cov1, cov2)
}
\arguments{
\item{mu1}{the mean vectors of distribution1}

\item{mu2}{the mean vectors of distribution2}

\item{cov1}{the covariance matrix of distribution1}

\item{cov2}{the covariance matrix of distribution2}
}
\value{
Wasserstain distance between two distributions
}
\description{
objective: compute wasserstain distance between two Gaussian distributions
}
